separate sentry config and middleware to prevent circular import
